Reimplementation of OptionView & FileView (#5)
