+++ USFMConverter/App.axaml.cs
@@ -21,6 +21,8 @@ namespace USFMConverter
                     DataContext = new ViewData()
                 };
             }
+            
+            
 
             base.OnFrameworkInitializationCompleted();
         }
