+++ USFMConverter/UI/MainWindow.axaml
@@ -2,13 +2,12 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+        mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="600"
         xmlns:screen="using:USFMConverter.UI.Pages"
         x:Class="USFMConverter.MainWindow"
         Title="USFMConverter"
         Background="AliceBlue">
-  <StackPanel>
-    <!-- <screen:HomeScreen Name="HomeScreen" FolderSelected="HomeScreen_FolderSelected"/> -->
-    <screen:ProjectDetailScreen Name="ProjectDetailScreen" IsVisible="False"/>
-  </StackPanel>
+    <StackPanel>
+        <screen:ProjectDetailScreen Name="ProjectDetailScreen" />
+    </StackPanel>
 </Window>
