+++ USFMConverter/UI/Pages/FileView.axaml
@@ -1,0 +1,91 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="600"
+             x:Class="USFMConverter.UI.Pages.FileView"
+             Name="FileView">
+
+    <UserControl.Styles>
+        <Style Selector="TextBlock.Title">
+            <Setter Property="FontSize" Value="20" />
+            <Setter Property="FontWeight" Value="Bold" />
+        </Style>
+        <Style Selector="TextBlock.SelectedText">
+            <Setter Property="FontWeight" Value="Medium" />
+            <Setter Property="Background" Value="White" />
+        </Style>
+        <Style Selector="Button.RemoveButton">
+            <Setter Property="Foreground" Value="#dd4c46" />
+            <Setter Property="Background" Value="White" />
+        </Style>
+    </UserControl.Styles>
+
+    <StackPanel Spacing="5" Margin="20" Grid.Column="0">
+        <StackPanel>
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="3*" />
+                    <ColumnDefinition Width="7*" />
+                </Grid.ColumnDefinitions>
+
+                <StackPanel Grid.Column="0" HorizontalAlignment="Center">
+                    <TextBlock HorizontalAlignment="Center" FontWeight="Bold" FontSize="25" Margin="10">Project Ready!</TextBlock>
+                    <TextBlock HorizontalAlignment="Center" Margin="10 5 10 5">Your project is ready to be converted.</TextBlock>
+                    <Button HorizontalAlignment="Center" BorderBrush="Black" BorderThickness="0.5" Margin="15"
+                            Padding="15" Background="#335cff" Foreground="White">
+                        Convert Project
+                    </Button>
+                </StackPanel>
+                
+                
+                <Border Name="DragDropArea" DragDrop.AllowDrop="True" Grid.Column="1" 
+                        Background="#edf1fa" BorderBrush="#edf1fa"
+                        BorderThickness="1" CornerRadius="10">
+                    
+
+                    <StackPanel Margin="10 20 10 20">
+                        <TextBlock VerticalAlignment="Center" FontWeight="Bold"
+                                   Margin="10" Foreground="#244f9d">
+                            Add Files
+                        </TextBlock>
+                        <TextBlock VerticalAlignment="Center" Margin="10 0 10 10">
+                            You may browse your computer for files or drag and drop them directly to the screen to import.
+                        </TextBlock>
+
+                        <Border Margin="10" HorizontalAlignment="Left" CornerRadius="5" BorderThickness="2"
+                                BorderBrush="#737f8f">
+                            <Button Foreground="#244f9d" Background="#edf1fa" HorizontalAlignment="Center"
+                                    VerticalAlignment="Center">
+                                Browse Files
+                            </Button>
+                        </Border>
+                    </StackPanel>
+                </Border>
+                
+            </Grid>
+            <!-- </Border> -->
+        </StackPanel>
+
+        <TextBlock Classes="Title" Margin="0 30 0 0">Project Files</TextBlock>
+        
+        <Grid>
+            <TextBlock Classes="SelectedText" Margin="0 10 0 0"
+                       Background="White" VerticalAlignment="Center"
+                       HorizontalAlignment="Left">
+                0 Selected
+            </TextBlock>
+
+            <Button Classes="RemoveButton" Margin="0 10 0 0"
+                    Background="White" BorderThickness="0"
+                    HorizontalAlignment="Right">
+                - Remove Files
+            </Button>
+        </Grid>
+
+        <ListBox Padding="20" Name="FilesListBox" Height="300"
+                 ScrollViewer.VerticalScrollBarVisibility="Visible"
+                 SelectionMode="Multiple" Items="{Binding Files, Mode=TwoWay}" />
+    </StackPanel>
+
+</UserControl>
