--- USFMConverter/UI/Pages/HomeScreen.axaml
@@ -1,29 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="USFMConverter.UI.Pages.HomeScreen">
-  <UserControl.Styles>
-    <Style Selector="TextBlock.DragText">
-      <Setter Property="FontSize" Value="20"/>
-    </Style>
-    <Style Selector="Button">
-      <Setter Property="VerticalContentAlignment" Value="Center"/>
-      <Setter Property="HorizontalContentAlignment" Value="Center"/>
-      <Setter Property="Background" Value="Blue"/>
-      <Setter Property="Padding" Value="80,20"/>
-      <Setter Property="Foreground" Value="White"/>
-    </Style>
-    <Style Selector="Button:pointerover /template/ ContentPresenter">
-      <Setter Property="Background" Value="Blue"/>
-    </Style>
-  </UserControl.Styles>
-  <StackPanel Name="DragDropArea" Spacing="20">
-    <Image Source="/UI/Assets/start_screen.png" MaxHeight="500"/>
-    <TextBlock HorizontalAlignment="Center" Classes="DragText">Drag and Drop a USFM folder to get started.</TextBlock>
-    <Button HorizontalAlignment="Center" Click="BrowseFiles_Click" Name="BrowseFiles">
-      Browse Files
-    </Button>
-  </StackPanel>
-</UserControl>
