+++ USFMConverter/UI/Pages/SuccessScreen.axaml
@@ -1,0 +1,15 @@
+
+
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="600"
+             x:Class="USFMConverter.UI.Pages.SuccessScreen">
+
+    <StackPanel>
+        <!-- <Image Source="../Assets/check.png" Width="100" Height="100" Stretch="None"/> -->
+
+    </StackPanel>
+
+</UserControl>
