+++ USFMConverter/UI/Pages/SuccessScreen.axaml.cs
@@ -1,0 +1,20 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace USFMConverter.UI.Pages
+{
+    public class SuccessScreen : UserControl
+    {
+        public SuccessScreen()
+        {
+            InitializeComponent();
+        }
+
+        private void InitializeComponent()
+        {
+            AvaloniaXamlLoader.Load(this);
+        }
+    }
+}
+
